corrigindo e adicionando algumas  coisas
