+++ Pessoa.py
@@ -1,9 +1,12 @@
 from abc import ABC, abstractmethod #para ser uma classe abstrata - não pode ser instaciada (não possui objetos próprios dela)
 import uuid #para atribuicao de ids as pessoas do nosso database
+import random
 
+rd = random.Random()
+rd.seed(123)
 class Pessoa(ABC):
     def __init__(self, id, nome: str, email: str):
-        self.__id = str(uuid.uuid4())
+        self.__id = str(uuid.UUID(int=rd.getrandbits(128), version=4))
         self._nome = nome
         self.email = email
